Handle asyncio.TimeoutError exceptions

+++ cogs/commands/tags.py
@@ -53,14 +53,14 @@ class Tags(commands.Cog):
         try:
             tname = await self.bot.wait_for("message", check=check)
         except asyncio.TimeoutError:
-            return await ctx.send("Command timed out")
+            return await ctx.send("Command timed out", ephemeral=True)
 
         await ctx.send("Enter the contents of the tag")
 
         try:
             contents = await self.bot.wait_for("message", check=check)
         except asyncio.TimeoutError:
-            return await ctx.send("Command timed out")
+            return await ctx.send("Command timed out", ephemeral=True)
 
         async with self.bot.tags_cxn.cursor() as cur:
             await cur.execute(
@@ -102,6 +102,9 @@ class Tags(commands.Cog):
         if not result:
             return await ctx.send(f"`{tag}`: no such tag")
             
+        if int(result[1]) != ctx.author.id or not ctx.author.guild_permissions.administrator:
+            return await ctx.send(f"Either you don't own tag `{tag}` or you lack administrator permissions to remove it otherwise")
+        
         async with self.bot.tags_cxn.cursor() as cur:
             await cur.execute(
                 "DELETE FROM tags WHERE tag = ? AND guild_id = ?",
